requirements.txt
torch
torchvision
torchaudio
streamlit
scikit-learn
matplotlib
seaborn
pillow
numpy
